Fix GUI module visibility and tests
